+++ src/irvine/oeis/a378/A378694.java
@@ -19,7 +19,7 @@ public class A378694 extends LambdaSequence {
       final int s = 1;
       final int t = 7;
       final int u = 6;
-      return Rationals.SINGLETON.sum(0, n, k -> new Q(Binomial.binomial(t * k + u * (n - k) + r, k)).multiply(Binomial.binomial(n + (s - 1) * k - 1, n - k)).divide(t * k + u * (n - k) + r)).multiply(r).num();
+      return Rationals.SINGLETON.sum(0, n, k -> new Q(Binomial.binomial((long) t * k + (long) u * (n - k) + r, k)).multiply(Binomial.binomial(n + (long) (s - 1) * k - 1, n - k)).divide((long) t * k + (long) u * (n - k) + r)).multiply(r).num();
     });
   }
 }
