+++ src/irvine/oeis/a380/A380774.java
@@ -21,7 +21,7 @@ public class A380774 extends LambdaSequence {
       final int s = 0;
       final int t = -2;
       final int u = 2;
-      return Rationals.SINGLETON.sum(0, n, k -> new Q(r * n + (s - r) * k + q).pow(k - 1).multiply(Binomial.binomial(r * u * n + ((s - r) * u + t) * k + q * u, n - k)).divide(Functions.FACTORIAL.z(k))).multiply(Functions.FACTORIAL.z(n).multiply(q)).num();
+      return Rationals.SINGLETON.sum(0, n, k -> new Q((long) r * n + (long) (s - r) * k + q).pow(k - 1).multiply(Binomial.binomial((long) r * u * n + ((s - r) * u + t) * k + q * u, n - k)).divide(Functions.FACTORIAL.z(k))).multiply(Functions.FACTORIAL.z(n).multiply(q)).num();
     });
   }
 }
